Refactor: moved scripts to /plugins/microbot
