+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/gabplugs/goldrush/GabulhasGoldRushConfig.java
@@ -1,4 +1,4 @@
-package net.runelite.client.plugins.gabplugs.goldrush;
+package net.runelite.client.plugins.microbot.gabplugs.goldrush;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
